src/main/java/CountClumps.java
/*
 * Copyright (c) 2022. Lorem ipsum dolor sit amet, consectetur adipiscing elit.
 * Morbi non lorem porttitor neque feugiat blandit. Ut vitae ipsum eget quam lacinia accumsan.
 * Etiam sed turpis ac ipsum condimentum fringilla. Maecenas magna.
 * Proin dapibus sapien vel ante. Aliquam erat volutpat. Pellentesque sagittis ligula eget metus.
 * Vestibulum commodo. Ut rhoncus gravida arcu.
 */

public class CountClumps {
    public int countClumps(int... nums) {
        int value = 0;
        int count = 0;
        for (int i = 0; i < nums.length - 1; i++) {

            if (nums[i] == nums[i + 1]) {
                int value2 = nums[i];
                if (value != value2)
                    count++;

            } else {
                value = nums[i];
            }
            value = nums[i];
        }
        return count;
    }
}
